backend/requirements.txt
asgiref==3.10.0
colorama==0.4.6
distlib==0.3.9
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
iniconfig==2.1.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
drf-spectacular==0.27.1
